chore: update shareOnTwitter util
